requirements.txt
flask
flask-cors
python-dotenv
pypdf
spacy
groq
scikit-learn
